Update README.md
Hash Table – In-Class Practice
In this class we have seen how to has on integer values when the value is the same as a key. In this practice today you will create a hash table to hold string values. The size of the table will be whatever the user enters. The key for each value will be the combined ASCII value of every character in the string. Your program should print out the table as seen in our examples. You can handle collision using either open addressing or chaining. Finally, you will create a driver that takes in four values.